Increase max latency of removable drive plugin to 28 days

See merge request briar/briar!1503

+++ bramble-android/src/main/java/org/briarproject/bramble/plugin/file/AndroidRemovableDrivePluginFactory.java
@@ -35,7 +35,7 @@ public class AndroidRemovableDrivePluginFactory implements
 	}
 
 	@Override
-	public int getMaxLatency() {
+	public long getMaxLatency() {
 		return MAX_LATENCY;
 	}
