+++ bramble-core/src/main/java/org/briarproject/bramble/plugin/file/TransportOutputStreamWriter.java
@@ -26,7 +26,7 @@ class TransportOutputStreamWriter implements TransportConnectionWriter {
 	}
 
 	@Override
-	public int getMaxLatency() {
+	public long getMaxLatency() {
 		return plugin.getMaxLatency();
 	}
